+++ crates/miden-objects/src/account/mod.rs
@@ -425,7 +425,9 @@ mod tests {
         },
         asset::{Asset, AssetVault, FungibleAsset, NonFungibleAsset},
         testing::{
-            account_id::ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE,
+            account_id::{
+                ACCOUNT_ID_PRIVATE_SENDER, ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE,
+            },
             storage::AccountStorageDeltaBuilder,
         },
     };
@@ -445,16 +447,22 @@ mod tests {
 
     #[test]
     fn test_serde_account_delta() {
+        let account_id = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
         let final_nonce = Felt::new(2);
         let asset_0 = FungibleAsset::mock(15);
         let asset_1 = NonFungibleAsset::mock(&[5, 5, 5]);
-        let storage_delta = AccountStorageDeltaBuilder::default()
+        let storage_delta = AccountStorageDeltaBuilder::new()
             .add_cleared_items([0])
             .add_updated_values([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .build()
             .unwrap();
-        let account_delta =
-            build_account_delta(vec![asset_1], vec![asset_0], final_nonce, storage_delta);
+        let account_delta = build_account_delta(
+            account_id,
+            vec![asset_1],
+            vec![asset_0],
+            final_nonce,
+            storage_delta,
+        );
 
         let serialized = account_delta.to_bytes();
         let deserialized = AccountDelta::read_from_bytes(&serialized).unwrap();
@@ -464,6 +472,7 @@ mod tests {
     #[test]
     fn valid_account_delta_is_correctly_applied() {
         // build account
+        let account_id = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
         let init_nonce = Felt::new(1);
         let asset_0 = FungibleAsset::mock(100);
         let asset_1 = NonFungibleAsset::mock(&[1, 2, 3]);
@@ -504,14 +513,19 @@ mod tests {
 
         // build account delta
         let final_nonce = Felt::new(2);
-        let storage_delta = AccountStorageDeltaBuilder::default()
+        let storage_delta = AccountStorageDeltaBuilder::new()
             .add_cleared_items([0])
             .add_updated_values([(1, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .add_updated_maps([(2, updated_map)])
             .build()
             .unwrap();
-        let account_delta =
-            build_account_delta(vec![asset_1], vec![asset_0], final_nonce, storage_delta);
+        let account_delta = build_account_delta(
+            account_id,
+            vec![asset_1],
+            vec![asset_0],
+            final_nonce,
+            storage_delta,
+        );
 
         // apply delta and create final_account
         account.apply_delta(&account_delta).unwrap();
@@ -534,18 +548,20 @@ mod tests {
     #[should_panic]
     fn valid_account_delta_with_unchanged_nonce() {
         // build account
+        let account_id = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
         let init_nonce = Felt::new(1);
         let asset = FungibleAsset::mock(110);
         let mut account =
             build_account(vec![asset], init_nonce, vec![StorageSlot::Value(Word::default())]);
 
         // build account delta
-        let storage_delta = AccountStorageDeltaBuilder::default()
+        let storage_delta = AccountStorageDeltaBuilder::new()
             .add_cleared_items([0])
             .add_updated_values([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .build()
             .unwrap();
-        let account_delta = build_account_delta(vec![], vec![asset], init_nonce, storage_delta);
+        let account_delta =
+            build_account_delta(account_id, vec![], vec![asset], init_nonce, storage_delta);
 
         // apply delta
         account.apply_delta(&account_delta).unwrap()
@@ -555,6 +571,7 @@ mod tests {
     #[should_panic]
     fn valid_account_delta_with_decremented_nonce() {
         // build account
+        let account_id = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
         let init_nonce = Felt::new(2);
         let asset = FungibleAsset::mock(100);
         let mut account =
@@ -562,12 +579,13 @@ mod tests {
 
         // build account delta
         let final_nonce = Felt::new(1);
-        let storage_delta = AccountStorageDeltaBuilder::default()
+        let storage_delta = AccountStorageDeltaBuilder::new()
             .add_cleared_items([0])
             .add_updated_values([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .build()
             .unwrap();
-        let account_delta = build_account_delta(vec![], vec![asset], final_nonce, storage_delta);
+        let account_delta =
+            build_account_delta(account_id, vec![], vec![asset], final_nonce, storage_delta);
 
         // apply delta
         account.apply_delta(&account_delta).unwrap()
@@ -576,6 +594,7 @@ mod tests {
     #[test]
     fn empty_account_delta_with_incremented_nonce() {
         // build account
+        let account_id = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
         let init_nonce = Felt::new(1);
         let word = [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)];
         let storage_slot = StorageSlot::Value(word);
@@ -584,7 +603,8 @@ mod tests {
         // build account delta
         let final_nonce = Felt::new(2);
         let account_delta = AccountDelta::new(
-            AccountStorageDelta::default(),
+            account_id,
+            AccountStorageDelta::new(),
             AccountVaultDelta::default(),
             Some(final_nonce),
         )
@@ -595,13 +615,14 @@ mod tests {
     }
 
     pub fn build_account_delta(
+        account_id: AccountId,
         added_assets: Vec<Asset>,
         removed_assets: Vec<Asset>,
         nonce: Felt,
         storage_delta: AccountStorageDelta,
     ) -> AccountDelta {
         let vault_delta = AccountVaultDelta::from_iters(added_assets, removed_assets);
-        AccountDelta::new(storage_delta, vault_delta, Some(nonce)).unwrap()
+        AccountDelta::new(account_id, storage_delta, vault_delta, Some(nonce)).unwrap()
     }
 
     pub fn build_account(assets: Vec<Asset>, nonce: Felt, slots: Vec<StorageSlot>) -> Account {