+++ crates/miden-objects/src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std]
+#![allow(clippy::uninlined_format_args)]
 
 #[macro_use]
 extern crate alloc;