+++ crates/miden-tx/src/host/mod.rs
@@ -125,7 +125,7 @@ impl<'store, 'auth, A: AdviceProvider> TransactionHost<'store, 'auth, A> {
             adv_provider,
             mast_store,
             scripts_mast_store,
-            account_delta: AccountDeltaTracker::new(&account.into()),
+            account_delta: AccountDeltaTracker::new(account),
             acct_procedure_index_map: proc_index_map,
             output_notes: BTreeMap::default(),
             authenticator,